perf(ui): move ELK layout into its own module and load the engine on demand
`EntityLineageLayoutUtils` mixed two concerns: five pure geometry helpers
(`centerNodePosition`, `getNodesBoundsReactFlow`, `getViewportForBoundsReactFlow`,
...) and two ELK-backed layout functions. Because `ELKUtil` was imported at
module scope, importing any of the pure helpers dragged elkjs — 1.37MB built.

`LineageControlButtons` imports only `centerNodePosition`, and `LineageProvider`
is statically imported by eight lineage components. That was enough to land
elkjs in the shared chunk which 476 route chunks statically import, so every
authenticated page downloaded and parsed the lineage layout engine. Glossary
pages paid for it while rendering with G6/antv-dagre, never calling ELK at all.

`getELKLayoutedElements` and `positionNodesUsingElk` now live in
`utils/Lineage/Layout/ElkLayoutUtils.ts`, alongside a named `loadElkLayout()`
lazy boundary. `EntityLineageLayoutUtils` no longer references ELK in any form,
so the pure helpers cannot re-acquire the dependency by accident.

A plain file split would not have been enough on its own: `LineageProvider` is
the sole caller of the ELK functions and is itself reachable from the shared
cluster, so the engine needs the dynamic boundary regardless. Making it a named
function in an ELK-specific module keeps that explicit rather than burying an
`await import()` in the middle of layout logic.

Measured on a production build:

  ELKUtil chunk importers   1 static -> 0 static, 1 dynamic
  glossary terms tab        12.57 MB -> 11.20 MB   (-1.37 MB, runtime)

Verified against a live backend: the glossary terms tab makes zero elk-related
requests and still renders, and /lineage fetches ELKUtil + vendor-elk on demand
and draws 83 nodes with no console errors.
