+++ openmetadata-ui/src/main/resources/ui/src/context/LineageProvider/LineageProvider.tsx
@@ -131,8 +131,8 @@ import {
 import {
   centerNodePosition,
   getViewportForLineageExport,
-  positionNodesUsingElk,
 } from '../../utils/EntityLineageLayoutUtils';
+import { positionNodesUsingElk } from '../../utils/Lineage/Layout/ElkLayoutUtils';
 import {
   createNodes,
   getConnectedNodesEdges,
